Minor corrections here and there, main method can now be used to calculate orbits.

+++ Keplerian/src/main/java/keplerian/physics/RigidBody.java
@@ -64,7 +64,7 @@ public class RigidBody {
      * @param angularVelocity Angular velocity of the object.
      * @param parent The celestial body the object is orbiting around.
      */
-    public RigidBody(int mass, Vector3d relPos, Vector3d relVel, Vector3d facing, Vector3d angularVelocity, CelestialBody parent)
+    public RigidBody(float mass, Vector3d relPos, Vector3d relVel, Vector3d facing, Vector3d angularVelocity, CelestialBody parent)
     {
         this.mass = mass;
         this.relPos = relPos;
@@ -84,7 +84,7 @@ public class RigidBody {
      * @param orbit The orbit the object is currently on.
      * @param parent The celestial body the object is orbiting around.
      */
-    public RigidBody(int mass, Vector3d facing, Vector3d angularVelocity, Orbit orbit, CelestialBody parent)
+    public RigidBody(float mass, Vector3d facing, Vector3d angularVelocity, Orbit orbit, CelestialBody parent)
     {
         this.mass = mass;
         this.facing = facing;
